refactor: DTO 경로 변경
