+++ backend/src/main/java/com/dajava/backend/domain/image/dto/ImageDimensions.java
@@ -1,4 +1,4 @@
-package com.dajava.backend.domain.image;
+package com.dajava.backend.domain.image.dto;
 
 public record ImageDimensions(
 	int pageWidth,
