+++ backend/src/main/java/com/dajava/backend/domain/image/service/pageCapture/LocalFileStorageService.java
@@ -19,7 +19,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.dajava.backend.domain.image.ImageDimensions;
+import com.dajava.backend.domain.image.dto.ImageDimensions;
 import com.dajava.backend.domain.image.dto.ImageSaveResponse;
 import com.dajava.backend.domain.image.exception.ImageException;
 import com.dajava.backend.domain.register.entity.PageCaptureData;
